- making build command composer changes aware

+++ src/Console/BuildCommand.php
@@ -10,7 +10,7 @@ use RecursiveIteratorIterator;
 
 class BuildCommand extends ConsoleCommand
 {
-    protected $acceptedArgs = ['-v'];
+    protected $acceptedArgs = ['-v', '-d'];
 
     const DIST_FOLDER = './dist/';
     const COMPOSER_INSTALL = 'composer install --no-ansi --no-dev --no-interaction --no-progress --no-scripts --optimize-autoloader';
@@ -42,15 +42,38 @@ class BuildCommand extends ConsoleCommand
 
     public function run()
     {
-        $this->logInfo('Cleaning dist Folder');
-        $this->refreshDistDir();
+        $vendorUpdated = true;
+        $forceCleanup = $this->getArg('-d', false);
+
+        if (!$forceCleanup) {
+            $this->logInfo('checking previous build');
+            $vendorUpdated = $this->areVendorUpdated();
+            $this->logInfo('are vendor updated? ' . ($vendorUpdated ? 'yes' : 'no'));
+            $this->logInfo('skipping composer install step');
+        }
+
+        if ($vendorUpdated || $forceCleanup) {
+            $this->logInfo('Cleaning dist Folder');
+            $this->refreshDistDir();
+        }
+
         $this->logInfo('Copying Project files');
         $this->copyProjectStructure();
-        $this->logInfo('Running composer');
-        $this->runInstall();
-        $this->logInfo('Cleaning vendor/');
-        $this->cleanVendor();
+
+        if ($vendorUpdated || $forceCleanup) {
+            $this->logInfo('Running composer');
+            $this->runInstall();
+            $this->logInfo('Cleaning vendor/');
+            $this->cleanVendor();
+        }
+
+
+        $this->logInfo("******************************************");
         $this->logInfo("Done, don't forget to check the .env file");
+        if(!$vendorUpdated){
+            $this->logInfo('you dont need to push the vendor/ folder');
+        }
+        $this->logInfo("******************************************");
     }
 
     private function refreshDistDir()
@@ -63,7 +86,9 @@ class BuildCommand extends ConsoleCommand
 
     private function rmrdir($folderName)
     {
-        exec('rm -rf ' . $folderName);
+        if (file_exists($folderName)) {
+            exec('rm -rf ' . $folderName);
+        }
     }
 
     private function copyProjectStructure()
@@ -75,22 +100,28 @@ class BuildCommand extends ConsoleCommand
         }
 
         foreach ($this->projectFolders as $folder) {
+            $this->rmrdir(self::DIST_FOLDER . $folder);
             $this->dircopy($folder, self::DIST_FOLDER . $folder);
         }
     }
 
     private function dircopy($target, $destination)
     {
-        mkdir($destination, 0755);
+        if (!is_dir($destination)) {
+            mkdir($destination, 0755);
+        }
+
         foreach (
             $iterator = new RecursiveIteratorIterator(
                 new RecursiveDirectoryIterator($target, RecursiveDirectoryIterator::SKIP_DOTS),
                 RecursiveIteratorIterator::SELF_FIRST) as $item
         ) {
-            if ($item->isDir()) {
-                mkdir($destination . DIRECTORY_SEPARATOR . $iterator->getSubPathName());
-            } else {
-                copy($item, $destination . DIRECTORY_SEPARATOR . $iterator->getSubPathName());
+            $this->logInfo('copying item: ' . $item);
+            $fullPath = $destination . DIRECTORY_SEPARATOR . $iterator->getSubPathName();
+            if ($item->isDir() && !file_exists($fullPath)) {
+                mkdir($fullPath);
+            } elseif (!$item->isDir()) {
+                copy($item, $fullPath);
             }
         }
     }
@@ -118,5 +149,14 @@ class BuildCommand extends ConsoleCommand
         $this->logInfo('file count after cleanup');
         $filesAfter = shell_exec('find ./dist -type f | wc -l');
         $this->logInfo($filesAfter);
+    }
+
+    protected function areVendorUpdated(): bool
+    {
+        if (is_dir(self::DIST_FOLDER) && is_dir(self::DIST_VENDOR) && file_exists(self::DIST_FOLDER . 'composer.json')) {
+            return md5(file_get_contents('./composer.json')) !== md5(file_get_contents(self::DIST_FOLDER . 'composer.json'));
+        }
+
+        return false;
     }
 }
